+++ .gitignore
@@ -53,3 +53,4 @@ cdk.context.json
 data/
 
 Brewfile.lock.json
+*.xml
