+++ jest.config.js
@@ -5,4 +5,13 @@ module.exports = {
   transform: {
     '^.+\\.tsx?$': 'ts-jest',
   },
+  reporters: [
+    'default',
+    [
+      'jest-junit',
+      {
+        outputDirectory: './target',
+      },
+    ],
+  ],
 };
