+++ package.json
@@ -46,6 +46,7 @@
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
     "jest": "^29.6.1",
+    "jest-junit": "^16.0.0",
     "prettier": "^3.2.4",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
