+++ test/stateless/stateless-pipeline.test.ts
@@ -3,7 +3,6 @@ import { Annotations, Match } from 'aws-cdk-lib/assertions';
 import { SynthesisMessage } from 'aws-cdk-lib/cx-api';
 import { AwsSolutionsChecks, NagSuppressions } from 'cdk-nag';
 import { StatelessPipelineStack } from '../../lib/pipeline/orcabus-stateless-pipeline-stack';
-import { StatefulPipelineStack } from '../../lib/pipeline/orcabus-stateful-pipeline-stack';
 
 function synthesisMessageToString(sm: SynthesisMessage): string {
   return `${sm.entry.data} [${sm.id}]`;
@@ -17,44 +16,6 @@ describe('cdk-nag-stateless-pipeline', () => {
   beforeEach(() => {
     app = new App({});
     stack = new StatelessPipelineStack(app, 'TestStack', {
-      env: {
-        account: '123456789',
-        region: 'ap-southeast-2',
-      },
-    });
-    Aspects.of(stack).add(new AwsSolutionsChecks());
-
-    NagSuppressions.addStackSuppressions(stack, [
-      { id: 'AwsSolutions-IAM4', reason: 'Allow CDK Pipeline' },
-      { id: 'AwsSolutions-IAM5', reason: 'Allow CDK Pipeline' },
-      { id: 'AwsSolutions-S1', reason: 'Allow CDK Pipeline' },
-      { id: 'AwsSolutions-KMS5', reason: 'Allow CDK Pipeline' },
-      { id: 'AwsSolutions-CB3', reason: 'Allow CDK Pipeline' },
-    ]);
-  });
-
-  test('cdk-nag AwsSolutions Pack errors', () => {
-    const errors = Annotations.fromStack(stack)
-      .findError('*', Match.stringLikeRegexp('AwsSolutions-.*'))
-      .map(synthesisMessageToString);
-    expect(errors).toHaveLength(0);
-  });
-
-  test('cdk-nag AwsSolutions Pack warnings', () => {
-    const warnings = Annotations.fromStack(stack)
-      .findWarning('*', Match.stringLikeRegexp('AwsSolutions-.*'))
-      .map(synthesisMessageToString);
-    expect(warnings).toHaveLength(0);
-  });
-});
-
-describe('cdk-nag-stateful-pipeline', () => {
-  let stack: StatefulPipelineStack;
-  let app: App;
-
-  beforeEach(() => {
-    app = new App({});
-    stack = new StatefulPipelineStack(app, 'TestStack', {
       env: {
         account: '123456789',
         region: 'ap-southeast-2',
